requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
matplotlib>=3.9.2
seaborn>=0.12.0
scikit-learn==1.5.1
pyyaml==6.0.1
python-dotenv==1.0.1
tqdm==4.66.4

# Existing API clients
anthropic>=0.25.0
openai>=1.30.0
requests>=2.31.0

# Dependencies for your dashboard
plotly==5.17.0
scipy==1.11.4